+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 CHANGELOG
 ====
 
+dev
+----
+
+ * css-components: Fixed broken popover components.
+ * ons-popover: Changed internal DOM structure.
+
 v2.0.6
 ----